requirements.txt
# Core Framework
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
Werkzeug>=3.0.0

# SOAP Client for Planif-Neige API
zeep>=4.2.0
lxml>=5.0.0

# Background Jobs
APScheduler>=3.10.0

# Email Service
resend>=2.0.0

# HTTP Requests
requests>=2.31.0

# Geospatial
shapely>=2.0.0

# Environment
python-dotenv>=1.0.0

# Production Server
gunicorn>=21.0.0

# Testing
pytest>=7.4.0
pytest-flask>=1.3.0
pytest-cov>=4.1.0

# Rate Limiting
Flask-Limiter>=3.5.0
